Move component to shared folder
